+++ src/pages/IssueList/components/IssueList.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import { CssBaseline, Grid, Container } from '@material-ui/core';
 import { Docs, Issue } from '@store/type';
-import AppbarShift from '../layout/AppbarShift';
-import IssueTable from './components/IssueTable';
+import AppbarShift from '../../layout/AppbarShift';
+import IssueTable from './IssueTable';
 
 interface Props {
   issues: Docs<Issue>;
